More Features
My sprite sheet system was reworked and had to adjust this program for it. The building area now allows for a bigger map. Building area can now be dragged around work on different areas, zooming was added to zoom in or out of the building area. Current texture is now shown on the menu tab.