New Layer Feature
Layers can be created as well and can be toggled to be shown or not shown. Layer also have a traits that can be toggled through which are: 'solid' (default), 'platform', 'decoration'. The solid trait means the tiles are a completely solid object and characters in platformer games cannot pass through them. The platform trait means that characters in platformer games can go through them, but can stand on top of them. The decoration trait means that the tile cannot be collided with at all.